requirements.txt
streamlit>=1.28.0
python-docx>=1.0.0
openpyxl>=3.1.0
pandas>=2.0.0
matplotlib>=3.7.0
google-generativeai>=0.3.0
reportlab>=4.0.0
Pillow>=10.0.0
